requirements.txt
langchain 
langchain-community 
llama-index
sentence-transformers
ollama
chromadb
praw
python-dotenv
streamlit